requirements.txt

django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-rest-framework==0.1.0
django-restframework==0.0.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
